requirements.txt
discord.py>=2.7.1
wavelink>=3.5.1
aiohttp>=3.13.5
python-dotenv>=1.2.2
openai>=2.32.0
sentence-transformers>=5.4.1
chromadb>=1.5.8